refactor(budgeting): clean architecture, i18n, logging, and error handling
- Create TransactionPage domain record to remove Spring Data dependency from domain
- Extract system prompt as shared @Bean (BudgetingConfig)
- Add i18n infrastructure (messages.properties, Messages component)
- Add TransactionPageMapper to decouple controller from Spring Page conversion
- Add @Slf4j logging to all use cases, controller, repository, and tool
- Remove @Transactional from repository, keep only on use cases
- Change delete(Transaction) to deleteById(UUID) for cleaner API
- Add amount validation in PersistTransactionTool
- Replace @RequestMapping with @PutMapping/@DeleteMapping for consistency
- Fix @ConditionalOnProperty to use explicit matchIfMissing
- Add EntityNotFoundException handler to global RestExceptionHandler
- Add RestExceptionHandlerTest for 404 coverage
- Convert TransactionOutput.valor from double to BigDecimal
- Update all tests to match new signatures

+++ src/main/java/com/samuelgularte/financeflow/budgeting/application/output/TransactionOutput.java
@@ -9,7 +9,7 @@ public record TransactionOutput(
         String id,
         String description,
         String category,
-        double valor,
+        BigDecimal valor,
         String createdAt
 ) {
     public static TransactionOutput from(Transaction transaction) {
@@ -17,7 +17,7 @@ public record TransactionOutput(
                 transaction.id().toString(),
                 transaction.description(),
                 transaction.category().name(),
-                BigDecimal.valueOf(transaction.amount(), 2).setScale(2, RoundingMode.HALF_UP).doubleValue(),
+                BigDecimal.valueOf(transaction.amount(), 2).setScale(2, RoundingMode.HALF_UP),
                 transaction.createdAt().toString()
         );
     }
